#101: 환율 페이지 dto 추가
