+++ src/main/java/com/ssafy/ssashinsa/heyfy/exchange/dto/exchangeRate/TuitionDto.java
@@ -1,21 +1,21 @@
-package com.ssafy.ssashinsa.heyfy.exchange.dto;
-
-import java.time.LocalDate;
+package com.ssafy.ssashinsa.heyfy.exchange.dto.exchangeRate;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PeriodDto {
+public class TuitionDto {
+    private PeriodDto period;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMdd")
-    private LocalDate start;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMdd")
-    private LocalDate end;
+    private LocalDate recommendedDate;
+    private String recommendationNote;
 }
+
